requirements.txt
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
et-xmlfile==1.1.0
filelock==3.16.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
identify==2.6.1
idna==3.10
nodeenv==1.9.1
numpy==2.1.1
oauth2client==4.1.3
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
Telethon==1.37.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
